+++ common/src/main/java/net/mehvahdjukaar/supplementaries/mixins/SkeletonMixin.java
@@ -45,7 +45,7 @@ public abstract class SkeletonMixin extends AbstractSkeleton implements IQuiverE
     protected void dropCustomDeathLoot(DamageSource damageSource, int looting, boolean hitByPlayer, CallbackInfo ci) {
         if (this.quiver != null && hitByPlayer) {
             ItemStack itemStack = this.quiver;
-            if (Math.max(this.random.nextFloat() - (float) looting * 0.02F, 0.0F) < quiverDropChance) {
+            if (Math.max(this.random.nextFloat() - looting * 0.02F, 0.0F) < quiverDropChance) {
                 this.spawnAtLocation(itemStack);
                 this.quiver = ItemStack.EMPTY;
             }
